[Feat] Login Checker 추가하여 Layout과 분리
