+++ pages/404.tsx
@@ -1,7 +1,7 @@
 import { useSetRecoilState } from 'recoil';
 import { useEffect } from 'react';
 import { errorState } from 'utils/recoil/error';
-import ErrorPage from 'components/ErrorPage';
+import ErrorPage from './error';
 
 const Error = () => {
   const setErrorMessage = useSetRecoilState(errorState);
